auction入札状況確認
